fix(campaigns): add ARIA tab semantics, use shared constants
Add role=tablist/tab/tabpanel, aria-selected, aria-controls,
aria-labelledby, roving tabindex, and arrow-key navigation to
campaign tabs. Move CampaignTab type and CAMPAIGN_TABS constant
to @lfx-one/shared.

LFXV2-2027
