+++ apps/lfx-one/src/app/modules/dashboards/campaigns/campaigns.component.ts
@@ -3,20 +3,8 @@
 
 import { Component, signal } from '@angular/core';
 
-type CampaignTab = 'planning' | 'implementation' | 'monitoring' | 'optimization';
-
-interface TabOption {
-  id: CampaignTab;
-  label: string;
-  icon: string;
-}
-
-const TABS: TabOption[] = [
-  { id: 'planning', label: 'Planning', icon: 'fa-light fa-clipboard-list' },
-  { id: 'implementation', label: 'Implementation', icon: 'fa-light fa-rocket' },
-  { id: 'monitoring', label: 'Monitoring', icon: 'fa-light fa-chart-line' },
-  { id: 'optimization', label: 'Optimization', icon: 'fa-light fa-wand-magic-sparkles' },
-];
+import { CAMPAIGN_TABS } from '@lfx-one/shared/constants';
+import type { CampaignTab } from '@lfx-one/shared/interfaces';
 
 @Component({
   selector: 'lfx-campaigns',
@@ -25,10 +13,31 @@ const TABS: TabOption[] = [
   styleUrl: './campaigns.component.scss',
 })
 export class CampaignsComponent {
-  protected readonly tabs = TABS;
+  protected readonly tabs = CAMPAIGN_TABS;
   protected readonly selectedTab = signal<CampaignTab>('planning');
 
   protected selectTab(tab: CampaignTab): void {
     this.selectedTab.set(tab);
+  }
+
+  protected onTabKeydown(event: KeyboardEvent, currentIndex: number): void {
+    let newIndex: number | null = null;
+
+    if (event.key === 'ArrowRight') {
+      newIndex = (currentIndex + 1) % this.tabs.length;
+    } else if (event.key === 'ArrowLeft') {
+      newIndex = (currentIndex - 1 + this.tabs.length) % this.tabs.length;
+    } else if (event.key === 'Home') {
+      newIndex = 0;
+    } else if (event.key === 'End') {
+      newIndex = this.tabs.length - 1;
+    }
+
+    if (newIndex !== null) {
+      event.preventDefault();
+      this.selectTab(this.tabs[newIndex].id);
+      const target = (event.target as HTMLElement).parentElement?.children[newIndex] as HTMLElement | undefined;
+      target?.focus();
+    }
   }
 }
